% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adipose.R
\docType{data}
\name{adipose}
\alias{adipose}
\title{Adipose tissue data}
\format{A data frame with 118 rows and 167 columns:
\describe{
...
}}
\source{
\url{https://www.metabolomicsworkbench.org/}
}
\usage{
adipose
}
\description{
Metabolomics data of adipose tissue samples
}
\keyword{datasets}
