% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLmm.R
\name{fitLmm}
\alias{fitLmm}
\title{Fit linear mixed-effects models (LMMs)}
\usage{
fitLmm(fix, random, data, start, end=NULL, auto=FALSE, major=NULL, pval=0.05, ...)
}
\arguments{
\item{fix}{list or vector; list or vector of fixed effects}

\item{random}{list or vector; list or vector of random effects}

\item{data}{dataframe; a dataframe contains metadata and measured variables}

\item{start}{integer; index of the first measured variable}

\item{end}{integer; index of the last measured variable, If not given, all variables are fitted}

\item{auto}{logical; automatically include only significant fixed effects based on chi-square test, \code{defaut = FALSE}}

\item{major}{list or vector; list or vector of fixed effects always be included, if provided, \code{auto} is \code{= TRUE}}

\item{pval}{numeric; value of significance level for chi-square test, \code{defaut = 0.05}}

\item{...}{other arguments of \code{lmer} function}
}
\value{
\code{fitLmm} returns a LMM for each measured variable. Model information is returned as an object of class \code{limmo},
with the following components:
\code{completeMod} = list of LMMs containing all given fixed and random effects

\code{updateMod} = list of LMMs containing fixed and random effects, if \code{auto = TRUE} or \code{major} is provided

\code{testRes} = list of outputs from chi-square tests

\code{fittedDat} = a dataframe of processed data

\code{rawDat} = a dataframe of original data

\code{dataIndex} = vector of indices of the first and last measured variable
}
\description{
Fit a linear mixed-effects model (LMM) to each measured variable (e.g. metabolite) via \code{lmer} function of \pkg{lme4}.
}
\details{
The function processes a given data matrix (e.g. metabolomics data) by fitting a LMM to each measured variable via \code{lmer} function of \pkg{\link{lme4}}.
Significance level of the fixed-effect term is assessed by chi-square test as implemented in \code{drop1}.
If \code{auto = FALSE}, all fixed effects are included in models.
If \code{auto = TRUE}, only significant fixed effects with \code{Pr(Chi) < pval}, are included in models.
If \code{major} is provided, \code{major}-fixed effects and significant fixed effects with \code{Pr(Chi) < pval} are included in models.
Random effect must be provided to form LMMs. The random-effect term can be in several forms as given in Table 2 of Bates et al. (2015).
}
\examples{
##Ex 1. LMMs containing all given fixed and random effects
#fitMet = fitLmm(fix=c('Sex','Age','BMI','Stage','Location','Tissue'), random='(1|Id)', data=adipose, start=10, end=14)
##Ex 2. LMMs containing only significant fixed effects and random effects
#fitMet = fitLmm(fix=c('Sex','Age','BMI','Stage','Location','Tissue'), random='(1|Id)', data=adipose, start=10, end=14, auto=TRUE)
##Ex 3. LMMs containing only significant fixed effects, Tissue effect and random effects
#fitMet = fitLmm(fix=c('Sex','Age','BMI','Stage','Location','Tissue'), random='(1|Id)', data=adipose, start=10, end=14, major='Tissue')
#structure(fitMet) #show limmo object
#summary(fitMet$updateMod[[1]]) #summarize outputs of fitting a LMM to variable X1
}
\references{
https://cran.r-project.org/web/packages/lme4/index.html

Bates, D., et al., Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 2015. 67(1): p. 1-48.
}
\seealso{
\code{\link{lmer}}, \code{\link{drop1}}, \code{\link{update}}, \code{\link{summary}}
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
