% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.limmo}
\alias{plot.limmo}
\title{Plot \code{limmo} object}
\usage{
plot.limmo(x, type, fix, ...)
}
\arguments{
\item{x}{\code{limmo} object from \code{fitLmm}}

\item{type}{character; type of plot. See Details}

\item{fix}{character; name of a fixed effect, if \code{type ='fitted'}}

\item{...}{other arguments of \code{plot} function}
}
\value{
plot
}
\description{
plot \code{limmo} object.
}
\details{
Types of plots include:

\code{residual} - residual and a normal Q-Q plot

\code{randeff} - random effects plot

\code{fixeff} - heatmaps of coefficients and significance levels

\code{fitted} - plots of measured variables before and after LMM fitting

\code{chisq} - html table of p-values from chi-square test

\code{coeff} - html table of fixed-effects coefficients
}
\examples{
#fitMet = fitLmm(fix=c('Sex','Age','BMI','Stage','Location','Tissue'), random='(1|Id)', data=adipose, start=10, end=14)
#plot(fitMet, type='residual')
}
\references{
https://cran.r-project.org/web/packages/lme4/index.html
}
\author{
Kwanjeera W \email{kwanjeera.wan@mahidol.ac.th}
}
